# docs/swagger.yaml
definitions:
...
  models.EventAnalyticsResponse:
    properties:
      available_seats:
        type: integer
      created_at:
        type: string
      event_id:
        type: string
      event_status:
        type: string
      event_title:
        type: string
      sales_status:
        type: string
      sold_seats:
        type: integer
      tier_count:
        type: integer
      tiers:
        items:
          $ref: '#/definitions/models.EventTierAnalytics'
        type: array
      total_revenue:
        type: number
      total_seats:
        type: integer
    type: object
...
  models.EventCancellationRequest:
    properties:
      reason:
        maxLength: 500
        minLength: 10
        type: string
    required:
    - reason
    type: object
...
  models.EventSalesControlRequest:
    properties:
      action:
        enum:
        - pause
        - resume
        - stop
        type: string
      reason:
        type: string
    required:
    - action
    type: object
...
  models.PayoutRequest:
    properties:
      admin_notes:
        type: string
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        description: Optional - can be for specific event
        type: string
      id:
        type: string
      organizer:
        $ref: '#/definitions/models.User'
      organizer_id:
        type: string
      processed_at:
        type: string
      processed_by:
        type: string
      request_number:
        type: string
      request_type:
        description: event_payout, bulk_payout
        type: string
      status:
        description: pending, approved, rejected, paid
        type: string
      updated_at:
        type: string
    type: object
...
  models.PayoutRequestUpdate:
    properties:
      admin_notes:
        type: string
      status:
        enum:
        - approved
        - rejected
        - paid
        type: string
    required:
    - status
    type: object
...
  utils.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      fields:
        description: For validation errors
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/utils.ErrorInfo'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A scalable REST API for event ticketing system with secure authentication
    and role-based access control (RBAC)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Timro Ticket API
  version: "1.0"
paths:
...
  /api/v1/admin/events/analytics:
    get:
      description: Get analytics for all events with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.EventAnalyticsResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all events analytics (Admin)
      tags:
      - Admin
...
  /api/v1/admin/payout-requests:
    get:
      description: Get all payout requests from all organizers
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - pending
        - approved
        - rejected
        - paid
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PayoutRequest'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all payout requests (Admin)
      tags:
      - Admin Payouts
  /api/v1/admin/payout-requests/{id}:
    put:
      consumes:
      - application/json
      description: Approve, reject, or mark payout request as paid
      parameters:
      - description: Payout Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PayoutRequestUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PayoutRequest'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Update payout request status (Admin)
      tags:
      - Admin Payouts
...
  /api/v1/organizer/events/{id}/analytics:
    get:
      description: Get comprehensive analytics for an event including tier breakdown
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.EventAnalyticsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get event analytics (Organizer)
      tags:
      - Organizer
  /api/v1/organizer/events/{id}/cancel:
    put:
      consumes:
      - application/json
      description: Cancel an event with reason
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Cancellation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EventCancellationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Event'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Cancel an event (Organizer)
      tags:
      - Organizer
  /api/v1/organizer/events/{id}/sales/control:
    put:
      consumes:
      - application/json
      description: Pause, resume, or stop event sales
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Sales control request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EventSalesControlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Event'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Control event sales (Organizer)
      tags:
      - Organizer
...
  /api/v1/organizer/events/{id}/tiers:
    post:
      consumes:
      - application/json
      description: Create a new pricing tier for an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Tier creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateEventTierRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.EventTier'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Create event tier (Organizer)
      tags:
      - Organizer
...
  /api/v1/organizer/payout-requests:
    get:
      description: Get all payout requests for the authenticated organizer
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - pending
        - approved
        - rejected
        - paid
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PayoutRequest'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get organizer payout requests (Organizer)
      tags:
      - Organizer
    post:
      consumes:
      - application/json
      description: Create a new payout request for earned revenue
      parameters:
      - description: Payout request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PayoutRequestCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PayoutRequest'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Create payout request (Organizer)
      tags:
      - Organizer
  /api/v1/organizer/payout-summary:
    get:
      description: Get payout summary including earnings, received amount, and pending
        requests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get payout summary (Organizer)
      tags:
      - Organizer
...
  /api/v1/organizer/tiers/{tierID}:
    delete:
      description: Delete a pricing tier (only if no tickets sold)
      parameters:
      - description: Tier ID
        in: path
        name: tierID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete event tier (Organizer)
      tags:
      - Organizer
    put:
      consumes:
      - application/json
      description: Update an existing pricing tier
      parameters:
      - description: Tier ID
        in: path
        name: tierID
        required: true
        type: string
      - description: Tier update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEventTierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.EventTier'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Update event tier (Organizer)
      tags:
      - Organizer
...
securityDefinitions:
  ApiKeyAuth:
    description: 'JWT token authentication. Use the ''Bearer'' prefix followed by
      a space and the access token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"